Fixing bug caused by having dash in organisation name
When we add an organisation with dash in its name to an audience then it causes the website to break. After reviewing, it turned out that the problem is caused by frontend rendering the dash. The problem was fixed byhaving replacing the '-' with the word 'dash' in the controller which passes the object to the html pages. And replacing the word 'dash' with '-' in the html page. This only affect the frontend so if other APIs use our api then this will not cause any problem.