LC-465-LC-469-LC-498
Built forntend pages for adidng a new question, quiz manage and generate reoporting page. All of them have been configures as well. So, when the link is clicked then the right page is loaded.

Delete main.css file

Delete main.css.map

LC-469: includes 'add new question' and 'add image' html templates

LC-469 updating adding question template to follow the new design

made different classes for each DTO objects
This commit also includes the changes for adding quiz when the user goes to the skills page

LC-466: adding profession specific logic

When a user go to the Manage quiz page, then there is a logic that checks if the user has set profession in their profile and if so, then it will load the skills quiz for that profession. If the profession is not set then it will show an error message worning the user to set the profession

LC466: fixed accessibility & PR comments

some of the pages were not accessible. So, I changed the design of the add question page to make it accessible with some styling changing. I also done the changes suggested by Kamesh on the PR request.

LC-466: added some checks

added logic to hide page if professions is not set and to display a text message when no questions is being added

LC-532: Adding description to a quiz

Made a template for the edit quiz description and make the node connection

added edit quiz

Changes are made for the skills Controller test class

Edit question - not complete

Treat answers as any other object