hotfix: adding additional conditions for agency capacity validation
Adding an additional client-side validation check on the capacity value to be less than or equal to the max boundary of java integers (i.e 2147483647).
Defined min max capacity validation will still be done by the api, but if we send a value outside the boundaries of a java int, we're limited in the ability to catch and provide custom error message.